more

+++ main.go
@@ -15,6 +15,59 @@ import (
 	yaml "gopkg.in/yaml.v2"
 )
 
+type state int
+
+const (
+	startState state = iota
+	deployappState
+	enabledrState
+	failoverState
+	relocateState
+	disabledrState
+	deleteappState
+	endState
+)
+
+func (s state) String() string {
+	return [...]string{
+		"start",
+		"deployapp",
+		"enabledr",
+		"failover",
+		"relocate",
+		"disabledr",
+		"deleteapp",
+		"end",
+	}[s]
+}
+
+func (s state) EnumIndex() int {
+	return int(s)
+}
+
+func getState(s string) state {
+	switch s {
+	case "start":
+		return startState
+	case "deployapp":
+		return deployappState
+	case "enabledr":
+		return enabledrState
+	case "failover":
+		return failoverState
+	case "relocate":
+		return relocateState
+	case "disabledr":
+		return disabledrState
+	case "deleteapp":
+		return deleteappState
+	case "end":
+		return endState
+	default:
+		return startState
+	}
+}
+
 // Global Variables
 var (
 	hub               string
@@ -31,52 +84,24 @@ var (
 	ramenOpsNamespace = "ramen-ops"
 	stopAt            string
 	startAt           string
+	startAtState      state
+	stopAtState       state
+	currentState      = startState
 )
+
+func showCurrentState() {
+	log.Printf("Current state: %s\n", currentState)
+}
 
 func main() {
 	setup()
-
-	if startAt == "start" {
-		deployApplications()
-	}
-	if stopAt == "deployapp" {
-		return
-	}
-
-	if startAt == "enabledr" || startAt == "start" {
-		enableDR()
-	}
-	if stopAt == "enabledr" {
-		return
-	}
-
-	if startAt == "failover" || startAt == "enabledr" || startAt == "start" {
-		failover()
-	}
-	if stopAt == "failover" {
-		return
-	}
-
-	if startAt == "relocate" || startAt == "failover" || startAt == "enabledr" || startAt == "start" {
-		relocate()
-	}
-	if stopAt == "relocate" {
-		return
-	}
-
-	if startAt == "disabledr" || startAt == "relocate" || startAt == "failover" || startAt == "enabledr" || startAt == "start" {
-		disableDR()
-	}
-	if stopAt == "disabledr" {
-		return
-	}
-
-	if startAt == "deleteapp" || startAt == "disabledr" || startAt == "relocate" || startAt == "failover" || startAt == "enabledr" || startAt == "start" {
-		deleteApplications()
-	}
-	if stopAt == "end" {
-		return
-	}
+	deployApplications()
+	enableDR()
+	failover()
+	relocate()
+	disableDR()
+	deleteApplications()
+	showCurrentState()
 }
 
 func setup() {
@@ -107,17 +132,19 @@ func setup() {
 
 	stopAt = os.Getenv("STOP_AT")
 	if stopAt == "" {
-		stopAt = "enabledr"
+		stopAt = "deployapp"
 	}
+	stopAtState = getState(stopAt)
 
 	startAt = os.Getenv("START_AT")
 	if startAt == "" {
 		startAt = "start"
 	}
+	startAtState = getState(startAt)
 
 	rbdCountEnv := os.Getenv("DEPLOYMENT_RBD_COUNT")
 	if rbdCountEnv == "" {
-		rbdCountEnv = "1"
+		rbdCountEnv = "0"
 	}
 	rbdCount, _ = strconv.Atoi(rbdCountEnv)
 
@@ -133,7 +160,6 @@ func setup() {
 	}
 	namespacesCount, _ = strconv.Atoi(namespacesCountEnv)
 
-	rand.Seed(time.Now().UnixNano())
 	// testID = fmt.Sprintf("test-%d", time.Now().UnixNano()) // Unique test ID
 	testID = fmt.Sprintf("test-%d", 1) // Unique test ID
 	namespaces = generateNamespaces(namespacesCount)
@@ -151,17 +177,29 @@ func setup() {
 	log.Printf("NAMESPACES: %v\n", namespaces)
 	log.Printf("STOP_AT: %s\n", stopAt)
 	log.Printf("START_AT: %s\n", startAt)
+	log.Printf("START_AT_STATE: %s\n", startAtState)
+	log.Printf("STOP_AT_STATE: %s\n", stopAtState)
 }
 
 func deployApplications() {
+	if startAtState > deployappState || stopAtState < deployappState {
+		return
+	}
+
 	createNamespaces(namespaces, preferredCluster)
+	createNamespaces(namespaces, failoverCluster)
 	// Deploy resources for RBD and CephFS storage classes
 	deployResources("rook-ceph-block", rbdCount, preferredCluster)
 	deployResources("rook-cephfs", cephfsCount, preferredCluster)
+	currentState = deployappState
 }
 
 func deleteApplications() {
-	deleteResources()
+	if startAtState > deleteappState || stopAtState < deleteappState {
+		return
+	}
+	deleteResources(preferredCluster)
+	deleteResources(failoverCluster)
 
 	for _, namespace := range namespaces {
 		deleteResource(preferredCluster, "namespace", namespace, "")
@@ -179,10 +217,24 @@ func deleteApplications() {
 		log.Fatalf("Failed to get pod name: %v", err)
 	}
 
-	cmd = exec.Command("kubectl", "--context", preferredCluster, "-n", "rook-ceph", "exec", "-it", strings.Trim(string(podName), "'"), "--", "bash", "-c", "rbd ls --pool=replicapool | xargs -n 1 rbd rm --pool=replicapool")
-	if err := cmd.Run(); err != nil {
-		log.Fatalf("Failed to delete rbd volumes: %v", err)
+	var rbdCount int
+	cmd = exec.Command("kubectl", "--context", preferredCluster, "-n", "rook-ceph", "exec", "-it", strings.Trim(string(podName), "'"), "--", "bash", "-c", "rbd ls --pool=replicapool | wc -l")
+	out, err := cmd.Output()
+	if err == nil {
+		rbdCount, _ = strconv.Atoi(strings.Trim(string(out), "\n"))
 	}
+	if rbdCount == 0 {
+		log.Printf("No rbd volumes to delete")
+	}
+
+	if rbdCount > 0 {
+		cmd = exec.Command("kubectl", "--context", preferredCluster, "-n", "rook-ceph", "exec", "-it", strings.Trim(string(podName), "'"), "--", "bash", "-c", "rbd ls --pool=replicapool | xargs -n 1 rbd rm --pool=replicapool")
+		if err := cmd.Run(); err != nil {
+			log.Printf("command failed: %s", cmd.String())
+			log.Fatalf("Failed to delete rbd volumes: %v", err)
+		}
+	}
+	currentState = deleteappState
 }
 
 func createNamespaces(namespaces []string, kcontext string) {
@@ -193,6 +245,9 @@ func createNamespaces(namespaces []string, kcontext string) {
 }
 
 func enableDR() {
+	if startAtState > enabledrState || stopAtState < enabledrState {
+		return
+	}
 	drpcYaml := loadTemplate("drpc.yaml", "", "", "")
 	placementYaml := loadTemplate("placement.yaml", "", "", "")
 	placementDecisionYaml := loadTemplate("placementdecision.yaml", "", "", "")
@@ -220,6 +275,7 @@ func enableDR() {
 	if err != nil {
 		log.Fatalf("Failed to patch resource status: %v", err)
 	}
+	currentState = enabledrState
 }
 
 // getResourceDetailsFromYAML returns the resource type, name and namespace of the kubernetes resource from a single resource yaml
@@ -245,36 +301,66 @@ func deleteResource(kcontext, restype, name, namespace string) {
 	var cmd *exec.Cmd
 
 	if namespace == "" {
-		cmd = exec.Command("kubectl", "delete", restype, name, "--context", kcontext, "--ignore-not-found")
+		cmd = exec.Command("kubectl", "delete", restype, name, "--context", kcontext, "--ignore-not-found", "--wait=false")
 	} else {
-		cmd = exec.Command("kubectl", "delete", restype, name, "-n", namespace, "--context", kcontext, "--ignore-not-found")
+		cmd = exec.Command("kubectl", "delete", restype, name, "-n", namespace, "--context", kcontext, "--ignore-not-found", "--wait=false")
 	}
 
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	if err := cmd.Run(); err != nil {
+		log.Printf("command failed: %s", cmd.String())
 		log.Fatalf("Failed to delete resource: %v", err)
 	}
 }
 
+func removeFinalizer(kcontext, restype, name, namespace string) {
+	time.Sleep(100 * time.Millisecond)
+
+	// check if resource exists, if not return
+	cmd := exec.Command("kubectl", "get", restype, name, "-n", namespace, "--context", kcontext)
+	if err := cmd.Run(); err != nil {
+		log.Printf("Resource not found, skipping finalizer removal")
+		return
+	}
+
+	cmd = exec.Command("kubectl", "patch", restype, name, "-n", namespace, "--type", "json", "-p", "[{\"op\": \"remove\", \"path\": \"/metadata/finalizers\"}]", "--context", kcontext)
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		log.Printf("command failed: %s", cmd.String())
+		log.Fatalf("Failed to remove finalizer: %v", err)
+	}
+}
+
 func disableDR() {
+	if startAtState > disabledrState || stopAtState < disabledrState {
+		return
+	}
 	drpcYaml := loadTemplate("drpc.yaml", "", "", "")
 	restype, name, ns := getResourceDetailsFromYAML(drpcYaml)
 	deleteResource(hub, restype, name, ns)
+	removeFinalizer(hub, restype, name, ns)
 
 	placementYaml := loadTemplate("placement.yaml", "", "", "")
 	restype, name, ns = getResourceDetailsFromYAML(placementYaml)
 	deleteResource(hub, restype, name, ns)
+	removeFinalizer(hub, restype, name, ns)
 
 	placementDecisionYaml := loadTemplate("placementdecision.yaml", "", "", "")
 	restype, name, ns = getResourceDetailsFromYAML(placementDecisionYaml)
 	deleteResource(hub, restype, name, ns)
+	removeFinalizer(hub, restype, name, ns)
 
 	deleteResource(hub, "manifestwork", fmt.Sprintf("%s-drpc-%s-ns-mw", testID, ramenOpsNamespace), preferredCluster)
 	deleteResource(hub, "manifestwork", fmt.Sprintf("%s-drpc-%s-ns-mw", testID, ramenOpsNamespace), failoverCluster)
+	currentState = disabledrState
 }
 
 func failover() {
+	if startAtState > failoverState || stopAtState < failoverState {
+		return
+	}
 	drpcYaml := loadTemplate("drpc.yaml", "", "", "")
 	applyYaml(drpcYaml, hub)
 
@@ -285,9 +371,13 @@ func failover() {
 	if err != nil {
 		log.Fatalf("Failed to patch resource status: %v", err)
 	}
+	currentState = failoverState
 }
 
 func relocate() {
+	if startAtState > relocateState || stopAtState < relocateState {
+		return
+	}
 	log.Default().Println("Relocating resources")
 
 	drpcYaml := loadTemplate("drpc.yaml", "", "", "")
@@ -300,6 +390,7 @@ func relocate() {
 	if err != nil {
 		log.Fatalf("Failed to patch resource status: %v", err)
 	}
+	currentState = relocateState
 }
 
 func generateNamespaces(count int) []string {
@@ -329,37 +420,42 @@ func createResources(storageClass string, count int) []string {
 	return resources
 }
 
-func deleteResources() {
-	for _, namespace := range namespaces {
-		cmd := exec.Command("kubectl", "delete", "deployment", "--all", "-n", namespace, "--context", preferredCluster, "--ignore-not-found")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			log.Fatalf("Failed to delete deployments: %v", err)
-		}
+func getResources(cluster, namespace, resourceType string) []string {
+	cmd := exec.Command("kubectl", "get", resourceType, "-n", namespace, "--context", cluster, "-o", "jsonpath='{.items[*].metadata.name}'")
+	out, err := cmd.Output()
+	if err != nil {
+		log.Fatalf("Failed to get %s: %v", resourceType, err)
+	}
+	resources := strings.Split(strings.Trim(string(out), "'"), " ")
 
-		cmd = exec.Command("kubectl", "delete", "pvc", "--all", "-n", namespace, "--context", preferredCluster, "--ignore-not-found")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			log.Fatalf("Failed to delete pvcs: %v", err)
+	filteredResources := []string{}
+	for _, name := range resources {
+		if name != "" {
+			filteredResources = append(filteredResources, name)
 		}
 	}
 
-	for _, namespace := range namespaces {
-		cmd := exec.Command("kubectl", "delete", "deployment", "--all", "-n", namespace, "--context", failoverCluster, "--ignore-not-found")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			log.Fatalf("Failed to delete deployments: %v", err)
-		}
+	return filteredResources
+}
 
-		cmd = exec.Command("kubectl", "delete", "pvc", "--all", "-n", namespace, "--context", failoverCluster, "--ignore-not-found")
-		cmd.Stdout = os.Stdout
-		cmd.Stderr = os.Stderr
-		if err := cmd.Run(); err != nil {
-			log.Fatalf("Failed to delete pvcs: %v", err)
-		}
+func forceDeleteResources(cluster, namespace, resourceType string) {
+	resources := getResources(cluster, namespace, resourceType)
+	for _, resource := range resources {
+		deleteResource(cluster, resourceType, resource, namespace)
+	}
+
+	time.Sleep(1 * time.Second)
+	resources = getResources(cluster, namespace, resourceType)
+	for _, resource := range resources {
+		removeFinalizer(cluster, resourceType, resource, namespace)
+	}
+}
+
+func deleteResources(cluster string) {
+	for _, namespace := range namespaces {
+		forceDeleteResources(cluster, namespace, "deployment")
+		forceDeleteResources(cluster, namespace, "pvc")
+		forceDeleteResources(cluster, namespace, "volumereplication")
 	}
 }
 
@@ -384,6 +480,12 @@ func loadTemplate(filename, deploymentName, storageClass, namespace string) stri
 	yamlContent = strings.ReplaceAll(yamlContent, "${NAMESPACE}", namespace)
 	yamlContent = strings.ReplaceAll(yamlContent, "${PREFERRED_CLUSTER}", preferredCluster)
 	yamlContent = strings.ReplaceAll(yamlContent, "${FAILOVER_CLUSTER}", failoverCluster)
+
+	if filename == "pvc.yaml" && storageClass == "rook-cephfs" {
+		yamlContent = strings.ReplaceAll(yamlContent, "${ACCESS_MODE}", "ReadWriteMany")
+	} else {
+		yamlContent = strings.ReplaceAll(yamlContent, "${ACCESS_MODE}", "ReadWriteOnce")
+	}
 
 	if filename == "drpc.yaml" {
 		yamlContent = strings.ReplaceAll(yamlContent, "${PROTECTED_NAMESPACES}", formatProtectedNamespaces(namespaces))
